+++ lib/plausible_web/controllers/billing_controller.ex
@@ -94,6 +94,8 @@ defmodule PlausibleWeb.BillingController do
     plan = Plausible.Billing.Plans.for_product_id(plan_id)
 
     if plan do
+      cycle = if plan[:monthly_product_id] == plan_id, do: "monthly", else: "yearly"
+      plan = Map.merge(plan, %{cycle: cycle, product_id: plan_id})
       usage = Plausible.Billing.usage(conn.assigns[:current_user])
 
       render(conn, "upgrade_to_plan.html",