+++ test/plausible_web/controllers/api/stats_controller/pages_test.exs
@@ -16,33 +16,37 @@ defmodule PlausibleWeb.Api.StatsController.PagesTest do
              ]
     end
 
-    test "calculates bounce rate for pages", %{conn: conn, site: site} do
+    test "calculates bounce rate and time on page for pages", %{conn: conn, site: site} do
       conn =
         get(
           conn,
-          "/api/stats/#{site.domain}/pages?period=day&date=2019-01-01&include=bounce_rate"
+          "/api/stats/#{site.domain}/pages?period=day&date=2019-01-01&detailed=true"
         )
 
       assert json_response(conn, 200) == [
                %{
+                 "time_on_page" => 82800,
                  "bounce_rate" => 33.0,
                  "count" => 3,
                  "pageviews" => 3,
                  "name" => "/"
                },
                %{
+                 "time_on_page" => 1,
                  "bounce_rate" => nil,
                  "count" => 2,
                  "pageviews" => 2,
                  "name" => "/register"
                },
                %{
+                 "time_on_page" => nil,
                  "bounce_rate" => nil,
                  "count" => 1,
                  "pageviews" => 1,
                  "name" => "/contact"
                },
                %{
+                 "time_on_page" => nil,
                  "bounce_rate" => nil,
                  "count" => 1,
                  "pageviews" => 1,