+++ test/plausible_web/controllers/unsubscribe_controller_test.exs
@@ -15,6 +15,13 @@ defmodule PlausibleWeb.UnsubscribeControllerTest do
       report = Repo.get_by(Plausible.Site.WeeklyReport, site_id: site.id)
       assert report.recipients == []
     end
+
+    test "renders success if site or weekly report does not exist in the database", %{conn: conn} do
+      conn =
+        get(conn, "/sites/nonexistent.com/weekly-report/unsubscribe?email=recipient@email.com")
+
+      assert html_response(conn, 200) =~ "Unsubscribe successful"
+    end
   end
 
   describe "GET /sites/:website/monthly-report/unsubscribe" do
@@ -29,6 +36,13 @@ defmodule PlausibleWeb.UnsubscribeControllerTest do
 
       report = Repo.get_by(Plausible.Site.MonthlyReport, site_id: site.id)
       assert report.recipients == []
+    end
+
+    test "renders success if site or weekly report does not exist in the database", %{conn: conn} do
+      conn =
+        get(conn, "/sites/nonexistent.com/monthly-report/unsubscribe?email=recipient@email.com")
+
+      assert html_response(conn, 200) =~ "Unsubscribe successful"
     end
   end
 end